Put back command using job as well
