+++ packages/twenty-server/src/database/commands/database-command.module.ts
@@ -6,11 +6,12 @@ import { DataSourceModule } from 'src/metadata/data-source/data-source.module';
 import { TypeORMModule } from 'src/database/typeorm/typeorm.module';
 import { WorkspaceModule } from 'src/core/workspace/workspace.module';
 import { DataSeedWorkspaceCommand } from 'src/database/commands/data-seed-dev-workspace.command';
-import { DataSeedDemoWorkspaceCronCommand } from 'src/database/commands/data-seed-demo-workspace.cron.command';
+import { DataSeedDemoWorkspaceCronCommand } from 'src/database/commands/crons/data-seed-demo-workspace.cron.command';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 import { WorkspaceSyncMetadataModule } from 'src/workspace/workspace-sync-metadata/workspace-sync-metadata.module';
 import { ObjectMetadataModule } from 'src/metadata/object-metadata/object-metadata.module';
 import { WorkspaceAddTotalCountCommand } from 'src/database/commands/workspace-add-total-count.command';
+import { DataSeedDemoWorkspaceCommand } from 'src/database/commands/data-seed-demo-workspace-command';
 
 @Module({
   imports: [
@@ -24,6 +25,7 @@ import { WorkspaceAddTotalCountCommand } from 'src/database/commands/workspace-a
   ],
   providers: [
     DataSeedWorkspaceCommand,
+    DataSeedDemoWorkspaceCommand,
     DataSeedDemoWorkspaceCronCommand,
     WorkspaceAddTotalCountCommand,
     ConfirmationQuestion,
